Initial commit of: -'main.py' -'input_processing.py' -'pdf_generation.py' -'grocery_list.json' -'readme.md'
Project is currently in a working state, but is not yet complete.
The goal is to create a 'master checklist' that is fully customizable.
Using a JSON file, the user can create a list of items that they want to
be able to check off. The program will then generate a PDF file with the
items listed in a table, and a checkbox next to each item. The user can
then print the PDF for their use case.